crates/textgrid-rs/scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pympi-ling>=1.71"]
